Fix masking for metrics
During evaluation, the masks of a model are not taken into account,
resulting in incorrect accuracies for sequence-to-sequence models with
padded output values. An example can be found in issue 4752.

To solve this issue, treat the metrics in a way analogous to the
objective functions, by adding an optional axis argument to metrics.py
and computing the overall mean only after multiplying with the model's
mask and dividing by its length in training.py. This required a small
reorganisation of the metric preparation in training.py.

Metric masking is now fixed for binary_accuracy, categorical_accuracy,
sparse_categorical_accuracy, top_k_categorical_accuracy,
mean_squared_error, mean_absolute_error, mean_absolute_percentage_error,
mean_squared_logarithmic_error, hinge, squared_hinge,
categorical_crossentropy, sparse_categorical_crossentropy,
binary_crossentropy, poisson, kullback_leibler_divergence and cosine_proximity.
Tests to confirm are included in tests/test_metric_masking.py.

For precision, recall, fbeta_score and fmeasure a slightly different
approach is followed, as they do not require taking a mean. In these
cases y_true and y_pred are multiplied with the mask before passing them
as arguments to the metric function.

A fix for matthews_correlation is not written yet. When a masked metric
should be computed for this function a warning is raised.